algorithms/mql5/tradingview_to_mt5_bridge/requirements.txt
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
